+++ DRN.Test.Unit/Tests/Framework/Testing/Providers/SettingsProviderTests.cs
@@ -1,7 +1,46 @@
+using DRN.Framework.SharedKernel.Domain;
+using DRN.Framework.SharedKernel.Enums;
+
 namespace DRN.Test.Unit.Tests.Framework.Testing.Providers;
 
 public class SettingsProviderTests
 {
+    [Fact]
+    public void Development_Should_Select_Partition_And_Preserve_Settings_Precedence()
+    {
+        using var defaults = SettingsProvider.Development();
+        defaults.NexusAppSettings.AppId.Should().Be(DefaultApp.AppId);
+        defaults.Configuration.GetValue<byte>("NexusAppSettings:AppId").Should().Be(DefaultApp.AppId);
+
+        using var settings = SettingsProvider.Development<TestApp>(
+            new { ApplicationName = "First", NexusAppSettings = new { AppInstanceId = 3 } },
+            new { ApplicationName = "Final", NexusAppSettings = new { AppId = TestApp.AppId, AppInstanceId = 7 } });
+        settings.Environment.Should().Be(AppEnvironment.Development);
+        settings.ApplicationName.Should().Be("Final");
+        settings.NexusAppSettings.AppId.Should().Be(TestApp.AppId);
+        settings.Configuration.GetValue<byte>("NexusAppSettings:AppId").Should().Be(TestApp.AppId);
+        settings.NexusAppSettings.AppInstanceId.Should().Be(7);
+        var key = settings.NexusAppSettings.GetDefaultKey();
+        settings.Dispose();
+        var readDisposedKey = () => { _ = key.MacKey.Bytes; };
+        readDisposedKey.Should().Throw<ObjectDisposedException>();
+    }
+
+    [Theory]
+    [DataInlineUnit((string?)null)]
+    [DataInlineUnit("")]
+    [DataInlineUnit(" ")]
+    [DataInlineUnit("5")]
+    [DataInlineUnit("128")]
+    [DataInlineUnit("invalid")]
+    public void Development_Should_Reject_Overrides_That_Do_Not_Match_Declared_AppId(string? appId)
+    {
+        var createDefault = () => SettingsProvider.Development(new { NexusAppSettings = new { AppId = appId } });
+        createDefault.Should().Throw<ConfigurationException>().WithMessage("*AppId must match the declared AppId 0*");
+        var createTyped = () => SettingsProvider.Development<DefaultApp>(new { NexusAppSettings = new { AppId = appId } });
+        createTyped.Should().Throw<ConfigurationException>().WithMessage("*AppId must match the declared AppId 0*");
+    }
+
     [Fact]
     public void SettingsProvider_Should_Return_IAppSettings_Instance()
     {